Add validation of certificates
I'm not completely happy with certvalidator as a dependency. It's a
sisterproject of asn1cryto that this project already uses, but both
seem rather unmaintained. However, I did not find a more suitable
implementation. pyca/cryptography seems to only implement low-level
primitives which would leave the implementation of the CRL or OCSP
fetching to us. This sounds just as error prone.